fix: restore file names
