+++ BioMain/include/BioI2C.h
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <ChNil.h>
 
-#include "BioParams.h"
+#include "Params.h"
 
 #if defined(I2C)
