+++ BioMain/include/ErrorThread.h
@@ -1,8 +1,8 @@
 #include <Arduino.h>
 #include <ChNil.h>
 
-//#include "BioParams.h"
-#include "hack/BioParams.h"
+//#include "Params.h"
+#include "hack/Params.h"
 
 #ifdef THR_ERROR
