+++ BioMain/include/StepsThread.h
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <ChNil.h>
 
-#include "hack/BioParams.h"
+#include "hack/Params.h"
 
 #ifdef THR_STEPS
