--- BioMain/include/hack/BioFunc.h
@@ -1,1 +1,0 @@
-void resetParameters();
