--- BioMain/include/hack/BioSem.h
@@ -1,5 +1,0 @@
-#include <ChNil.h>
-
-extern semaphore_t lockTimeCriticalZone;
-// extern SEMAPHORE_DECL(lockTimeCriticalZone, 1); // only one process in some
-// specific zones
