+++ BioMain/include/hack/EEPROMLogger.cpp
@@ -10,8 +10,8 @@
 
 #ifdef THR_EEPROM_LOGGER
 
-#include "BioParams.h"
 #include "Params.h"
+#include "Funcs.h"
 
 unsigned long lastLog = millis();
 uint16_t eepromLoggerID = 0;
