+++ BioMain/include/hack/Funcs.cpp
@@ -1,3 +1,6 @@
+#include "Params.h"
+#include "Funcs.h"
+
 #include <Arduino.h>
 #include <ChNil.h>
 //#include <TimeLib.h>
@@ -5,8 +8,8 @@
 
 #include "libraries/time/TimeLib.h"
 
-#include "BioParams.h"
-#include "EEPROMHack.h"
+#include "Params.h"
+#include "HackEEPROM.h"
 #include "SSTLogger.h"
 #include "ToHex.h"
 
@@ -143,3 +146,36 @@ uint8_t printCompactParameters(Print* output, uint8_t number) {
 uint8_t printCompactParameters(Print* output) {
   return printCompactParameters(output, MAX_PARAM);
 }
+
+void resetParameters() {
+  setAndSaveParameter(PARAM_STEPPER_SPEED, 60);
+  setAndSaveParameter(PARAM_STEPPER_SECONDS, 30);
+  setAndSaveParameter(PARAM_STEPPER_WAIT, 5);
+
+  setAndSaveParameter(PARAM_TEMP_EXT1, ERROR_VALUE);
+  setAndSaveParameter(PARAM_TEMP_EXT2, ERROR_VALUE);
+  setAndSaveParameter(PARAM_TEMP_PCB, ERROR_VALUE);
+  setAndSaveParameter(PARAM_PID, 0);
+  setAndSaveParameter(PARAM_TEMP_TARGET, 3000);
+
+  setAndSaveParameter(PARAM_WEIGHT_MAX, -32768);
+  setAndSaveParameter(PARAM_WEIGHT_FACTOR, 0);
+  setAndSaveParameter(PARAM_WEIGHT_EMPTY, 0);
+
+  setAndSaveParameter(PARAM_CURRENT_STEP, 0);
+  setAndSaveParameter(PARAM_CURRENT_WAIT_TIME, 0);
+  for (uint8_t i = FIRST_STEP_PARAMETER; i <= LAST_STEP_PARAMETER; i++) {
+    setAndSaveParameter(i, 0);
+  }
+
+  //int active = 1 << FLAG_PID_CONTROL | 1 << FLAG_STEPPER_CONTROL | 1 << FLAG_FOOD_CONTROL | 1 << FLAG_PH_CONTROL | 1 << FLAG_GAS_CONTROL | 1 << FLAG_SEDIMENTATION | 1 << FLAG_RELAY_FILLING | 1 << FLAG_RELAY_EMPTYING | 1 << FLAG_PH_CALIBRATE | 1 << FLAG_RELAY_ACID | 1 << FLAG_RELAY_BASE;
+
+  int active = 1 << FLAG_PID_CONTROL | 1 << FLAG_STEPPER_CONTROL | 1 << FLAG_OUTPUT_1 | 1 << FLAG_OUTPUT_2 | 1 << FLAG_OUTPUT_3 | 1 << FLAG_OUTPUT_4 | 1 << FLAG_PH_CONTROL | 1 << FLAG_GAS_CONTROL | 1 << FLAG_SEDIMENTATION | 1 << FLAG_RELAY_FILLING | 1 << FLAG_RELAY_EMPTYING | 1 << FLAG_PH_CALIBRATE | 1 << FLAG_RELAY_ACID | 1 << FLAG_RELAY_BASE;
+
+  int enable = 1 << FLAG_PID_CONTROL | 1 << FLAG_STEPPER_CONTROL | 1 << FLAG_OUTPUT_1 | 1 << FLAG_OUTPUT_2 | 1 << FLAG_OUTPUT_3 | 1 << FLAG_OUTPUT_4;
+
+  setAndSaveParameter(PARAM_STATUS, active);
+  setAndSaveParameter(PARAM_ENABLED, enable);
+  setAndSaveParameter(PARAM_ERROR, 0);
+}
+
