+++ BioMain/include/hack/Serial.h
@@ -4,7 +4,7 @@
 
 #include <Arduino.h>
 #include <ChNil.h>
-#include "BioParams.h"
+#include "Params.h"
 
 #ifdef THR_SERIAL
