+++ BioMain/include/hack/Start.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <ChNil.h>
 
-#include "Params.h"
+#include "Funcs.h"
 
 void checkParameters();
 void initParameters();
