+++ BioMain/include/hack/Weight.cpp
@@ -1,12 +1,12 @@
 #include <Arduino.h>
 #include <ChNil.h>
 
-#include "BioParams.h"
 #include "Params.h"
+#include "Funcs.h"
 
 #ifdef THR_WEIGHT
 
-#include "BioSem.h"
+#include "Sem.h"
 // SEMAPHORE_DECL(lockTimeCriticalZone, 1); // only one process in some specific
 // zones
