+++ BioMain/include/hack/WireMasterCh.h
@@ -1,4 +1,4 @@
-#include "BioParams.h"
+#include "Params.h"
 
 // setting ATmega32U4 as I2C slave.
 #ifdef THR_WIRE_SLAVE
