+++ BioMain/include/hack/WireSlaveCh.h
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <ChNil.h>
 
-#include <BioParams.h>
+#include <Params.h>
 
 // setting ATmega32U4 as I2C slave.
 #ifdef THR_WIRE_SLAVE
