+++ BioMain/src/OutputsThread.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <ChNil.h>
 
-#include "BioParams.h"
+#include "Params.h"
 #include "Status.h"
 
 #if defined(THR_OUTPUTS)
