+++ BioMain/src/SSTThread.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <ChNil.h>
 
-#include "BioParams.h"
+#include "Params.h"
 
 #ifdef THR_SST_LOGGER
