+++ BioMain/src/SerialThread.cpp
@@ -1,6 +1,6 @@
 #include <Arduino.h>
 #include <ChNil.h>
-#include "BioParams.h"
+#include "Params.h"
 
 #ifdef THR_SERIAL
